refactor: 이미지 코드 수정
